Add top5 endpoints for events and serviceproducts

+++ src/main/java/com/example/eventplanner/repositories/event/EventRepository.java
@@ -1,7 +1,7 @@
 package com.example.eventplanner.repositories.event;
 
-import com.example.model.event.Event;
+import com.example.eventplanner.model.event.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
