+++ src/main/java/com/example/eventplanner/repositories/serviceproduct/ServiceProductRepository.java
@@ -1,8 +1,7 @@
 package com.example.eventplanner.repositories.serviceproduct;
 
-import com.example.model.event.Event;
-import com.example.model.serviceproduct.ServiceProduct;
+import com.example.eventplanner.model.serviceproduct.ServiceProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceProductRepository extends JpaRepository<ServiceProduct, Integer> {
+public interface ServiceProductRepository extends JpaRepository<ServiceProduct, Long> {
 }
